Create Basketball-Pop-A-Shot
Created in python on the Raspberry Pi as a beginners project. The tutorial I used was from "RaspberryPiTutorial's" channel on Youtube. It was my first introduction and application using the python language. I followed the tutorial but also made a couple modification through Adobe Photoshop in OS X and the overall concept of the game. This was a chance for me to experiment and understand more of the syntax and abilities of the python language. Hope you enjoy! To be able to play make sure you have Python 2.7.5 and pygame. Simply compile and play. The images for the background, cursor, and basketball can be downloaded below:

http://tinypic.com/view.php?pic=2mcagjm&s=5 (rename to background.png)
http://tinypic.com/view.php?pic=10wnsjr&s=5 (rename to crosshairsmouse.png)
http://tinypic.com/view.php?pic=34znf9g&s=5 (rename to pi.png)

*Remember to place all three images in the same file. Otherwise the game won't run.